Log error when emulator process fails
When an emulator test runs on a machine that doesn't support emulation,
the only thing that prints out is "Time out while waiting for line
matching [...]", which isn't particularly helpful. Given that many of us
have VMs that do not support nested virtualization, I figured it would
be better to print a descriptive message essentially saying "use a
different machine".

Bug: N/A
Test: existing emulator tests
Change-Id: Ied2018dc61ef4aa892e1d2951af1677bac79c9c4