Autoupdate of 3 submodules
*tools/adt/idea updated to b1d1ad6d4b7df93f1765b100b061fdef1f1df91f
added sync issue for removed APIs.

added code to translate to IDE sync issues.

Test: Existing
Bug: 244224699
Change-Id: I1a6d51ab566019d89e17a195e39dd77ca49dee3f

*tools/analytics-library updated to 31a7a1fcbea714f8c4b2abfcc6457afb68e3543e
resync'ed file with google3 version.

Test: Existing
Bug: 244224699
Change-Id: Ie3e56277a6db53c181fa11e8dfb50b5c23a5b7fd

*tools/base updated to d382517026b026f15ac147c7ed3b709a0eb12f78
Disable public registerTransform API.

This is the first CL for multiple CLs aimed at removing support
for Transform API.

This CL only generates an exception when code invokes the
'registerTransform' API. However, the API is still there and
associated interfaces are retained so we do not fail sync
but instead generate a sync issue (documented so we can follow
impacted users).

Test: removed all tests using the public API except JarJarTest
Bug: 244224699
Change-Id: I14ae44b4765b017e69409b76a4fe8b18e57e124c