[SDK Index] E2E test for policy issues
This change adds a new test to replace manual testing for policy
issues. The test sets true to the flag that causes policy issues to be
shown and then verifies that those issues are shown. This change also
modifies SdkIndexPsdTest to verify that the default flag setting causes
the issue to be logged but not shown in PSD.

The snapshot.gz was modified to include only the set of issues needed by
the tests (including a policy issue), as opposed to use a public
available snapshot.

Bug: 233789469
Test: PolicyIssuesPsdTest, SdkIndexPsdTest
Change-Id: Ib0fab4ac050087b9da1ac55cbd035f2caecf5c41