Extract creation of Native debug process from CidrCommandLineState and CidrLauncher
This is needed to make creation of Native debug process independent of ExecutionEnvironment and RunConfiguration that we don't have when we attach debugger to existing process.

CidrCommandLineState.startDebugProcess was only invoking CidrLauncher.startDebugProcess, so CidrCommandLineState as a wrapper could be deleted. CidrLauncher.startDebugProcess was invoked ovveride
CidrLauncher.createDebugProcess. This CidrLauncher.createDebugProcess I extracted to method ConnectLLDBTask.createDebugProcess and invoke now directly.

Bug: 229101714
Test: run everything Native/Dual/ attach/ start
Change-Id: Ie2fac15d6cbe550832edd3bfb3606eb089f29538