Encapsulate build parameters in a separate object
Before this change GradleBuildInvoker.Request contained both:
1) fields describing the gradle build, like tasks, project path,
jvm arguments, etc.
And
2) Setup related to current run from studio, like if should wait for
completion, additional listener, project object, etc.

First group describes the build and can be saved and restored to repeat
the build. We need to store it as part of result of Build Analyzer.
Second group belongs solely to current build run an should not be copied
over or stored.
In order to assist with it, I encapsulate first group of parameters to a
separate object that can be stored, recovered and re-used later.
I keep the existing API so nothing should change in existing code.

Bug: N/A
Test: Existing tests should cover, no functionality changed
Change-Id: Ib4788199a59f4e200f61522190803dc3ae5396fc