Fix compilation on windows
Original Report : https://github.com/google/perfetto/issues/347

Note: This fix doesn't address all the issues reported by the original
github issue. It only fix the issue with posix-shared-memory.

As of this CL, perfetto works well on windows with default gn args.

However it doesn't work with `enable_perfetto_ipc=false` as reported in
the same github bug. This CL doesn't fix the perfetto on windows when
this flag is set to false. As it is unrelated issue, it will be covered
in another CL.

Bug: b/245122401

Change-Id: I64df1c5bf418802a6b632c25ca0e7ff0c1546a7b