[BA] Add missing data to results interface
BuildAnalysisResults object contains all the data for the result
including build request and finished timestamp. However, these fields
were not part of the BuildEventsAnalysisResult interface, and thus in
all code interacting with it should have
received these fields separately.

We might not need this intermediate interface at all now, but it is
convenient to have it in order to easily override parts of data in
tests. So for now just add these fields to the interface and change
all the code (mostly tests) that uses it.

Bug: N/A
Test: Tests updated
Change-Id: Ie6df370ac1993d9b0afdd426c40378602b115f4d