Autoupdate of 2 submodules
*tools/adt/idea updated to 96b380efad15289088cdf085d3f4f62540fe3312
Add package name for debuggables in Transport

... which is required when a process name is different from
the package name. It's not the most common setup but it
does happen. The exact package name is needed to setup
JVMTI agent and establish the communication between the
daemon and the agent.

Bug: 230871525
Test: updated
Change-Id: Ide234a80daa6dce947cb77b28471cfac34c49844

*tools/base updated to 87897a757694af1bb27acfaa68d1c531a5526c1e
Add package name in Transport Pipeline

... which is required when a process name is different from
the package name. It's not the most common setup but it
does happen. The exact package name is needed to setup
JVMTI agent and establish the communication between the
daemon and the agent.

Bug: 230871525
Test: see companion change Ide234a80daa6dce947cb77b28471cfac34c49844
Change-Id: I9687cfdb7c6d7075ddc41e8a41f7832560bfb86a