Autoupdate of 2 submodules
*tools/adt/idea updated to 8ff1bd413002f8556391943754b8b438359f36e2
Update command handler for "shell,v2"

The "fake agent" "shell,v2" device command handler supports
a single command. This change ensures other commands are dispatched
to the default "shell,v2" handler.

Bug: 188067646
Test: Existing
Change-Id: I98189b67808e45b0bb16b638e044c47ce390ec9d

*tools/base updated to d0b6719323f01db0b4705522d5b6be5ce6987d39
adblib: Use shell,v2 protocol if available

When retrieving device properties, we don't have access to the
`shellCommand` extension because the `DeviceProperties`
class is at a lower level layer.

A previous change was made to detect older `shell`
implementations that use `\r\n' for newlines (instead of '\n').

This change ensure we use the `shell,v2` protocol if supported,
to avoid the problem entirely (and be more efficient as a side
effect).

Bug: 188067646
Test: Existing
Change-Id: I901fe795f862ed5ee818572e8e7eb74e2329c0b1