Prevent reparsing when artifact already selected
Re-selecting an artifact would cause the profiler
to reparse it. To prevent this unecessary computation,
the selection behavior is prevented if the user is
already viewing the respective artifact.

Added behavior where post capture, the resulting artifact's
implicit selection by the UI is registered as the last selection.
Other implicit selections registered are the init of a new ongoing
session, or the importation of session(s).

Bug: 159323345
Test: Updated
Change-Id: I29c815306418e7511b7ffef95dee54f6c1da894d