Autoupdate of 3 submodules
*prebuilts/tools updated to efe64423b537a9166fb5730d287c8b828542c2c0
Update core-ktx to 1.8.0 and activity-compose to 1.5.1

This is to support upcoming updates of the Compose and Wear templates.

Test: TemplateTest
Bug: N/A
Change-Id: I0efae6cc0bb8613a82e6ca52845d8262bd7c7a45

*tools/adt/idea updated to e7c431017f84c6ef8ac1ff0aad904959683375ee
Update core-ktx to 1.8.0 and activity-compose to 1.5.1

This is to support upcoming updates of the Compose and Wear templates.

Test: TemplateTest
Bug: N/A
Change-Id: I236759948c83bb5988d85e892191fd056b73fa7e

*tools/base updated to e47c849663c9b5ea15a43f1ba3948277ffa4a3bc
Update core-ktx to 1.8.0 and activity-compose to 1.5.1

This is to support upcoming updates of the Compose and Wear templates.

Test: TemplateTest
Bug: N/A
Change-Id: Idc8994e724502942a020e186712ab8853c5af18f