Autoupdate of 2 submodules
*tools/adt/idea updated to 7e0bdbd67e74f9106072001bdcbf947bfc7292d5
Moves the updated assets from the old location

This updates the assets that are actually used
for the installer.

Fixes: 195528343
Test: N/A
Change-Id: Ie0a8d4ad0cca9f15ad851ffdfaed8551e175b4f9

*tools/vendor/google updated to 3236ad217e5eaa366071e2d7acd7308b7bc65c51
Move updated assets to the right location

These assets have been updated, but are not
in the right location as they are not used.
Deleting them to avoid confusion and moving them
to the right location.

Note this entire directory may be unused,
but I'm not sure if adrt is using some of it.

Will update that separately.

Bug: 195528343
Test: N/A This will be moved to rapid
Change-Id: I8f678e91b0c9eb43788cd09937ece051951af46f