Complete cancelled Visual Linting tasks with exception
If the future is cancelled and is then queried, it throws a
CancellationException, but with the stack trace from when it was
cancelled, not when it was queried. That is not helpful to figure out
where is the cancelled future accessed from.
Instead of a cancellation, call completeExceptionally on the future.
That way, when it is later accessed, we should have a more relevant
stack trace.

Bug: 245608056
Test: N/A
Change-Id: I46188d96282c439222a851f5c7ac9c3165e1ed54