Avoid collapsing children when loading
When performing the "synchronize" operation, we reload the selected
directory's children, and show a "... loading ..." node. Previously,
this node was added by mutating the parent tree node and firing
nodeStructureChanged to trigger the update. However, this causes any
existing child nodes to be collapsed.

Instead, use the insertNodeInto method to add the node, which fires
a treeNodesInserted event, which doesn't require as much work.

Updated the test to stop expecting the nodeStructureChanged event,
and to expect the treeNodesInserted instead, but only on the initial
update; some tests use checkMockViewActiveDevice to check for changes
that do not necessarily trigger this event.

Fixes: 141321904
Test: updated
Change-Id: I9ab5a5dafa7a232057db98aa2f2ad0e90480b48c