Allow Extract Resource for complex Kotlin strings
Kotlin strings are parsed as a KtStringTemplateExpression element. When
the string is "simple" with no placeholders or otherwise special
characters, that element has a single KtLiteralStringTemplateEntry
child. When the string contains escaped characters, each escaped
character is represented by a KtEscapeStringTemplateEntry element, and
the containing expression will have multiple children.

Current logic only allows extracting a resource for strings made up of
singular KtLiteralStringTemplateEntry elements. But since we can also
resolve KtEscapeStringTemplateEntry statically, I'm modifying the logic
of the intention to allow those as well (or more accurately, to allow
the parent KtStringTemplateExpression of both those items, whenever
its children are all one of the known types).

Fixes: 184565311
Test: KotlinAndroidAddStringResourceTest
Change-Id: I6d9d026776735ee00dda838c917503a9196e4d18