[AUA] Support change in buildFeatures.buildConfig default
The processor tests for the project actually using BuildConfig in a
meaningful way, rather than just for the presence of
BuildConfig-related directives in build files.  This means that
testing the processor requires a substantially heavier project
model than our usual unit-like tests; substantive tests have been
added as integration tests using the OldAgpTest machinery.

Strictly speaking the BuildConfig source files are generated by build,
rather than sync.  This implies that the Upgrade Assistant should
block the upgrade if the project has never been built (or the
generated files have been recently cleaned), with the user prompted to
run the generated files task for any (potentially) affected modules.

The readMoreUrl for this processor currently points at a non-existing
redirect (/r/tools/upgrade-assistant/build-config-default) on DAC.
This should be fixed before this processor is shown to non-dev users.

Adjust the main integration tests so that this processor does not
block execution, by providing a generated file.  (We cannot simply
disable BuildConfig generation, as we support versions of AGP earlier
than the setting is available.)

Bug: 243130943
Bug: 245397495
Test: added
Change-Id: I302f71f4f334bf78910b2500d59213e5e0cdccff