Autoupdate of 2 submodules
*tools/adt/idea updated to df3c10bf968ff5ce36b3b2117987b4216fb07452
Update template test with non-root activity

b/200894865 involves a bug with adding activities that aren't in the
root package. Update test to cover that case.

Bug: 200894865
Test: TemplateTest.kt
Change-Id: I5e5149cd5fbbd5679798ea4ef504b7b8941283a8

*tools/base updated to c62978bfcdea0a22e5a6a2d4183f106269154e17
Generate correct databinding package import

All viewbinding classes are generated under the root project package with `databinding` appended. But when an activity (or other recipe) is created outside of the project's root package, the import statement was incorrectly pointing towards the sub-package rather than the root package.

This change updates all the calls to this import statement to use the
root application package instead.

The "interesting" part of the change is in viewBindingUtils.kt. The
rest of the change is effectively just passing various parameters
around.

Test: TemplateTest.kt (in separate linked cl)
Fixes: 200894865
Change-Id: Ie11bff7c99c5c9b15fbfa581f76e48891f179776