Update D8 R8 to 8.0.5-dev
This is a development snapshot, it's fine to use for studio canary
build, but not for BETA or release, for those we would need a release
version of R8 binaries. This build IS suitable for preview release
but IS NOT suitable for public release.

Built here: go/r8-releases/raw/8.0.5-dev
Test: ./gradlew check
Bug:

Change-Id: I1be66cadb503422cbb2ad5fd26844b8feac39831