Autoupdate of 2 submodules
*tools/adt/idea updated to af8874fe54816aa112cb730e0ccc8a16ba17fe9c
Fix flaky test in DeviceViewPanelTest

The main problem was that we did not wait for the
UPDATE_SCREENSHOT_TYPE_COMMAND to finish. And sometimes the command
finished after the animation of going into 3D mode.

Also the timing of the animation was based on real time.
Replaced that with a time that can be controlled in tests.
(Updated Toggle3dActionTest to use the test time as well.)

Fixes: 246371105
Test: DeviceViewPanelTest
Change-Id: I9dcf66f20dce497148894c83a46b50c147c7b46c

*tools/base updated to dbf1a20bd4c5ba4d44de59d78cc477bc89ac9c00
Add a getCurrentTimeMillis() to VirtualTimeScheduler

Bug: N/A
Test: N/A
Change-Id: I4c902b924370168d7bc0bb134ab98350cfd7b95b