Allow users to wait as long as they want while connecting
The gathering of compose nodes in large compose apps seem to take
a long time. Do not force a timeout, instead show a warning about
a slow connect, and the user can choose to either continue waiting
or disconnect.

When disconnecting from an app inspection client, we display the
text: "Dump Views" instead of "Disconnect" since we will still
fallback to the legacy client.

Bug: 246219349
Test: Added unit tests on monitor.
Change-Id: I773d3fb33c9141e85a114d7737bdca887356e495