Delete the old DSL
This version of the DSL focused on collecting input and reducing an
upgrade expression to find the next set of inputs to prompt for.
However, this doesn't fit the limitations of the current assistant
framework. Therefore, the DSL is being replaced by version focused on
constructing a static description object that can be compiled into the
assistant framework's data model. The old DSL's deletion and the new
DSL's addition are split across CLs for diff clarity.

Bug: 243983260
Test: N/A
Change-Id: I33f79becd860d60c3c64d6077116bce7024292b7