Autoupdate of 3 submodules
*tools/adt/idea updated to 45afbd26d88bb0d9fc67956549f276e1bea01652
set up ASfP project

Change-Id: I659b09108d27dba5d7e2417a566c58cb8d5226a4

*tools/base updated to 7dce8c6fcee044a70d226ff6c95c3fc71051a370
add ASfP to lnzipper visibility list

Change-Id: If95402e5e9301c20c94e024656c173f400d9d104

*tools/vendor/google updated to c65fa57f55d51e78aab17d970615edb80fb7468e
set up ASfP with AIDEGen poc

This CL initializes the code directory for ASfP and adds an inaugural
"feature" of syncing project via an action in the main menu. Currently
just a proof-of-concept, it directly invokes AIDEGen in a subprocess,
passing as input a list of modules derived from the project.

Test: Launch ASfP and trigger AIDEGen. Check logs for expected output
and observe changes within the IDE.

Change-Id: I456f76bd169dd0fd0e95406f0484ef7b15e05577