Pure type safe builder DSL describing upgrades
Builder style DSL for describing the ordering of steps in an upgrade.
Currently, there's only one kind of Step: a Raw list of Elements.
Elements are currently opaque until embeddings come in a later CL.
Steps also have a label that will populate the assistant's step header.

Steps can be ordered in sequence or in parallel and these orderings can
be arbitrarily nested.

Fixes: 243983260
Test: N/A
Change-Id: I36e03844cd6e14faebd409ec747a1e1db21105de