Autoupdate of 2 submodules
*tools/adt/idea updated to 0f14530578c27b5bbdfa984147d4c26a2600d312
Adapt to ddmlib API change

Change Ia6e9bbb0eb5c85d4e72a48bf1fe4f29f94a6995d
slightly updates the legacy view APIs used by Layout Inspector
to decouple interface and implementation.

This change updates production code to use the new API instead
of the deprecate ones, but also change test code that was mocking
ddmlib behavior using class types that are not valid anymore.

Bug: 241411829
Bug: 188067646
Test: Existing
Change-Id: Ie19ab3d3de85c94aa5ceff778c7f627baaa8ea85

*tools/base updated to d511cfd49649b88a5e35b28d6b778e6cd120dc46
ddmlib: Make DebugViewDumpHandler more abstract

The DebugViewDumpHandler surface area was too large, specifically
exposed `ClientImpl` (an implementation detail), making it difficult
to implement the `Client` interface.

For example `listViewRoots(DebugViewDumpHandler callback)`
expects callers to implement a `DebugViewDumpHandler`,
but in practice, the only thing callers are expected to provide
is a Ddms Chunk Type that are interested in, and a callback
to invoke when the ddms packet reply payload is available
(as a ByteBuffer). So there is no need for callers to be
exposed to `ClientImpl` implementation details.

This change is required to be able to provide a
non-ddmlib implementation of `Client` via the
`ClientManager` extension point.

Bug: 241411829
Bug: 188067646
Test: Existing
Change-Id: Ia6e9bbb0eb5c85d4e72a48bf1fe4f29f94a6995d