Autoupdate of 2 submodules
*tools/adt/idea updated to a7d213ba602cc0b17b1a2c1673aa595c4c3b8273
Add isPrimary and description fields to TaskCategory

This way we can assure that all primary categories will have
descriptions to be shown on the IDE side.

Fixes: 245303698
Test: updated tests
Change-Id: I2b881edb1119a1999a585c62d2300ea58a25639e

*tools/base updated to 9e0a071c58f8505761f625efe9f318ef1a533ae8
Add isPrimary and description fields to TaskCategory

This way we can assure that all primary categories will have
descriptions to be shown on the IDE side.

Fixes: 245303698
Test: updated tests
Change-Id: Ic8f0906c61033acbc2922b3366fff9e837230461