Autoupdate of 2 submodules
*tools/base updated to 9b596c98ddbac3b974691b4762684f47fe1c7624
Make it an error for NDK symlinking to fail

If the user set ndk.symlinkdir to something but we weren't able
to create the symlink then issue an error rather than a warning.

Bug: 240198042
Test: Existing
Change-Id: I0b2e186d3ba91fd0b4c7d607855df4ade47afb35

*prebuilts/studio/buildbenchmarks updated to 92aea340b205a03aa62e8bc03cf9c60d3806e8fa
Shorten NDK symlink path

The symlinking logic already appends 'ndk' to the path so wasn't
necessary to explicitly add it to the path.

Bug: 240198042
Test: Is Test
Change-Id: I1a91286fab2f3b4d87378990d866eb76beecff19