Autoupdate of 2 submodules
*tools/adt/idea updated to 2f6d19a6efb6a7d6ef1ba4cf3680e5cdfe957a16
Add a validation to check buildFeatures.aidl property

moduleData.aidlDir is assumed as non-null at the moment.
But it's not going to be true after ag/19995810 lands.

We need to add TemplateConstraint.Aidl and evaluate if the existing
project has the buildFeatures.aidl is set to true by building the
GradleModel from the existing project.

Bug: 247809666
Test: TemplateTest, Manual testing
Change-Id: I913e0444bf84d4541b78e34e96a1f52a7c3af55f

*tools/base updated to 09e6e43591ea9d05dc89f0434b2c44cff862887e
Add TemplateConstraint.Aidl

After ag/19995810 lands, AIDL directory needs to be created on the
condition buildFeatures.aidl is set to true.

This CL adds a constraint to check if the existing module sets
buildFeatures.aidl to true for the template that has
TemplateConstraint.Aidl constraint.

Change-Id: I35b4b996637937f6b38712240872c01734788244

Bug: 247809666
Test: TemplateTest, Manual test by creating a Aidl file from the context
menu

Change-Id: I17f147989e6ff43830ad6d323139b9384039c9da