Autoupdate of 3 submodules
*tools/adt/idea updated to 10e9944f4a27ab956cdb97e7204d478c1832d8d7
Update model files now that AIDL is false by default

This change also modifies AndroidGradleTestUtils to allow tests to
specify whether to include AIDL source files (false by default).
This functionality is needed for the tests in
AidlDefaultRefactoringProcessorTest.

This change keeps the version in
AidlDefaultRefactoringProcessor.necessity() set to 8.0.0-alpha04
because that will be the version of AGP in which AIDL will be
disabled by default.

Bug: 243130937
Test: existing
Change-Id: I4d56b911622229df623d257e9c5a440a843a1496

*tools/base updated to 4d8657625c1c58630752205137c1cc0bfe329db6
Make android.defaults.buildfeatures.aidl false by default

Bug: 243130937
Test: existing
Change-Id: Iba151cccd775d17029d42142a7be42f918e803df

*tools/vendor/google updated to c846e58e2ad45a75d5514f4d963036f2de7cd298
Update model files now that AIDL is disabled by default

Bug: 243130937
Test: existing
Change-Id: Ic95eff4aa193a760aaef27a206cb2e1b296bd8a7