Autoupdate of 2 submodules
*tools/adt/idea updated to 2fb072780522f615435eb609fed757726abad0c3
Adapt to removal of disable_live_allocations

This CL adapts the studio-side implementation
to the removal of the disable_live_allocations
field in the UserOptions proto.

Bug: none
Test: updated
Change-Id: I5778384d0422fb198550b715403e16e412efa517

*tools/base updated to 5484d3f6a5ba16978a519de3035cf2822be92f5a
Remove disable_live_allocation from user options

The disable_live_allocation field in the
UserOptions proto (child message of
CpuTraceConfiguration proto) is not used.
Therefore, we can remove it.

Bug: none
Test: existing
Change-Id: Ic839d058d094309e18f7dd7b1a54f21f1441faea