Temporarily remove ads-sdk-app
The ads-sdk-app benchmarks are currently causing issues due with the
dependency on the android-TiramisuPrivacySandbox. The revision of android-TPS is now currently out of date and contains api breaking changes. For these reasons the benchmark will be temporarily removed until the new version of the preview platform and benchmark project is checked in.

Bug: 242831966
Test: adk-sdk-app benchmarks removed
Change-Id: I430008e96fe4395c8f6ebe42ae571d244f85a8db