Autoupdate of 2 submodules
*tools/adt/idea updated to 18cbbc6108ef2cdc46fff2ddca392cb5061ea533
Don't assert that GradleVersion is deserializable

There should be no need to serialize or deserialize it using the
Ion-based serializer.  (If there are users of this, they will end up
being broken in forthcoming bug fixes to the GradleVersion class, so
let's find out as soon as we can...)

Bug: 242691473
Test: existing
Change-Id: I4de2efaf6a24c9321492dd9914e89ea2d86294c5

*tools/base updated to 495f2a24ab20e6ca3899b8ed2f05b25990eecca7
Remove unused constructor

It might once have been used to serialize IdeAndroidProject, but our
models no longer use GradleVersion.

Bug: 242691473
Test: existing
Change-Id: Ibbe3a3bdb7ecca371718c5043ef6c311b2b82a52