Fix overreporting of running thread status
When comparing the running thread status time of a trace
upload in studio vs perfetto, the time in studio was
magnitudes larger than perfetto. This was because the
states were stored by their start time only. Therefore
the last state would continue to be so even after the
tracing was actually done and would continue until
user-dictate stoppage of capture.

In this CL we fix this by appending a fake state to the
threads events. Specifically, a NO_ACTIVITY thread state
is added under two conditions. The original last state is
not NO_ACTIVITY already, and theres no (non-fake) states
already present. This effectively terminates the last thread
state, and thus gives accurate running thread status times.

Note: there is still a negligible difference between the
running state duration in perfetto vs studio, with perfetto
always being about .5 microseconds longer. However, this is
due to perfetto's precision of timestamps being in nanoseconds
while on studio-side it uses microseconds.

Bug: 215749159
Test: updated and added
Change-Id: Ife8285d4f2cc71f80d06d306d9013e7253e1564d