Fix logic of stopping apps for Native debugger
Only one debugger should be responsible for stopping an app if needed.
It will be the Native debugger, because it is always presented in comparison with Java.

New logic:
When a user stops a Java Debug process, we stop a Native Debug process
When a user, or we stop a Native Debug process, if it was launchMode(we started an app) we finish an Android process.

Fixes: 250627477
Test: manually
Change-Id: I03873860f3984ecfd1669c39fb331e4c9790959d