Autoupdate of 2 submodules
*tools/adt/idea updated to fb4366f804413040fe80387e30673abb985e1489
Start using AgpVersion where appropriate

Convert many uses of GradleVersion to use the new AgpVersion class
where the surrounding context is clearly working with AGP (e.g. the
version of the plugin applied to a project in Sync; the current
version used to build in the Build Analyzer; the current declared
version in the Upgrade Assistant...)

Not all places have been updated (e.g. in the AGP Upgrade Assistant we
still have plenty of bare GradleVersion types).  Follow-up changes
will be made to specific areas.  A separate follow-up change will
separate GradleVersion from AgpVersion (as it is almost certainly an
error to compare instances of the two classes) before addressing
correctness issues in parsing and comparison.

Bug: 242691473
Test: existing tests updated
Change-Id: I24ab5bd4b8252fa9ebbfbdd6f17ed7b6e6009314

*tools/base updated to 2d9067a6f77a10aecca9f7d3428dda2c6a615563
Define AgpVersion class

Many (though far from all) uses of GradleVersion are in fact
specialized to computations on versions of the Android Gradle Plugin,
for which we have substantially more constraints: we always have
numeric major, minor and micro versions, and any preview information
is appended after those.

Define a specialized class, for now mutually compatible in its
operations with GradleVersion, to represent versions specifically of
AGP.  In the near future the two classes will be decoupled and no
longer mutually-compatible, once all specifically-AGP versions use the
new class.

Bug: 242691473
Test: in accompanying and next CL
Change-Id: I8ef7105fb45824d92b2a8eca4dab730a9f714062