Autoupdate of 3 submodules
*prebuilts/tools updated to 831204f98455ee77e47dbd1b13e88a01c199e838
Add GSON library to tools/adt/idea

Needed for unittests

Bug: 230631228
Test: No behavior change
Change-Id: I4b92527ad1bbde9170cc486c26b68af0ebc00a5b

*tools/adt/idea updated to 780a1124dd9d825e977923a3f4d229f26f23dd90
Add GSON library to tools/adt/idea

Needed for unittests

Bug: 230631228
Test: No behavior change
Change-Id: Idf9fba0e9fa73b481d790c5807642a7b0d5608f9

*tools/base updated to ab4fc91f57a02d70be551ab0154d610f26d70a06
Move Ninja parser to com.android.utils

This is part of a code shuffle in support of ASfP (Android Studio
for Platform). Moving to the new location allows the code to be
accessed from the eventual ASfP plugin.

The reference prototype requiring this change: ag/19098964

Bug: 230631228
Test: No behavior change
Change-Id: Ifd43a9de22fb0c2ceb73bdafaafdcec0cb81c8de