Autoupdate of 2 submodules
*tools/adt/idea updated to 527b3c7adf153ad75ee1a4677ef57e30a7bd3633
Adapt to TraceType being moved under UserOptions

This change simply adapts the studio-side code
to the move of the TraceType enum to under the
UserOptions proto.

Bug: none
Test: updated
Change-Id: I4e447ba624dcab74de80baf8658d764d8b71adc6

*tools/base updated to 6b36115ae132534f27b8f36070d18c2eed63d8f8
Move TraceType to under UserOptions

Because it is only used for UserOptions, the enum
defining different trace types of "TraceType" is
now nested/scoped under the UserOptions proto.

To avoid errors in our legacy code, the same TraceType
definition was used under the only other proto using
TraceType: CpuProfilingAppStopRequest.

Bug: none
Test: existing
Change-Id: I859c4834edbc48023dcd1b4cea0fc9265aae6350