Refactor ChangeEvents.
ChangeEvent being a sealed class requires all subclasses to live in the
same file which is getting too big. Additionally, the main benefit of it
being a sealed class is not used in the code base, namely exhaustive
`when` clause, instead we rely on the polymorphic transition method in
each subclass.

So this changes converts `ChangeEvent` to an interface and moves each
change into its own file additionally moving action dispatch to a
separate `actions` submodule.

bug: na
Test: existing tests.
Change-Id: I4ae47134c385bf8dcc6b2d7c9b95b96eeaacfa42