Fix NlPropertiesModel sharing MergingQueue
This ensures that when models share one queue, they still schedule
updates separately. Without this, some updates would be lost. This would
happen for example with the MotionEditor and the Attributes panel where
the first update for the latter would be lost because they are sharing
one queue and the updates would happen very close to each other.

Fixes: 247726011
Test: NlPropertiesModelTest
Change-Id: Iefad3b345939ed15e93505929c806c8ae10cf542