Autoupdate of 3 submodules
*tools/base updated to 75b023bb68f4fade890e50c8926c1a60a7774130
Copy release prebuilts to artifacts directory

This change copies the additional data needed by the Rapid release
workflow to the artifacts directory so that they can be downloaded.

Bug: 234630178
Test: N/A
Change-Id: Ie48755ecf695e4aa6d3ed1f971bedb0a0c74dab0

*tools/vendor/google updated to 95887f524a9712950e981967519d7dc1391155c4
Package cros-skeleton and nsis as zips

cros-skeleton and nsis are used to create ChromeOS and Windows
installers, respectively. These files are needed by the Rapid release
workflow, so this change packages both directories as zip files.

These files are kept in their original locations to retain compatibility
with ADRT.

Bug: 234630178
Test: N/A
Change-Id: Ic517894c8d32988313c70be434875b57123e29a6

*tools/vendor/intel updated to 1212619833f94165022ac6ba46a434d88fef37b4
Package Intel HAXM binaries in zip

The Intel HAXM binaries are needed to generate the Windows installer.
This change packages the binaries in a zip so that they can be used by
the Rapid workflow.

Bug: 234630178
Test: N/A
Change-Id: I501a8ea1a1df350d195aa596a2488c35998cbce1