Fix margin attribute handling for RelativeLayout
When adding a layout_margin attribute for a RelativeLayout, we want to
replace it with individual layout_margin_... attributes for each side,
and remove the layout_margin one.
But we need to wait to remove the layout_margin until all the others
have been added and committed, otherwise we may end up writing some
intermediate value and not the full value of layout_margin.

Fixes: 212943795
Test: checked manually
Change-Id: I5e22649315645c1b36404eb33b866e1e91ee7775