PairedDevicesPanel.pairingDeviceRemoved can be called in any thread
PairedDevicesPanel implements PairingStatusChangedListener and was
marking pairingDeviceRemoved as @UiThread. It does not make sense for an
implementation of a listener to try to force a specific thread so,
instead, this change makes PairedDevicesPanel to handle the case where
the listener does not happen in the UI thread (the default case).

Fixes: 251175683
Test: Additional tests in PairedDevicesPanelTest
Change-Id: I5dfb1abb4c76f72d717cc89907ede591573685b4