Autoupdate of 2 submodules
*tools/adt/idea updated to fd44143f335f0dd29dd7649d821acf3a6f3d2df6
Remove the UsesMinSdkAttributes lint check

This CL removes the IDE inspection registration for the UsesMinSdkAttributes
lint check (which was removed in lint).

Test: N/A
Bug: Not filed
Change-Id: I78eb5200da626ef627cd4de94ce9932e99128bd9

*tools/base updated to 5d4c1b30945f286f69b55ed1c4ba2af33b5dec38
Remove the UsesMinSdkAttributes lint check

This is one of the original lint checks from 2011, and makes sure that
you specify a minSdkVersion as well as a targetSdkVersion in the
manifest, which users sometimes forgot.

With Gradle, this is no longer the right place to set it; you set these
values in build.gradle via the DSL and then Gradle will inject it at
build time, so we stopped reporting these from Gradle a long time ago.
They often have false positives in other build-system integrations as
well, and do not really provide any real world value anymore, so this CL
just deletes it. (The issue is also added to the deletedIssues property
of the IssueRegistry such that lint will gracefully accept references to
this issue id in lint.xml files, baselines, command line flags, etc.

This lint check was used in a bunch of unit tests (as a simple issue to
trigger), so this CL also replaces these test usages with the
MultipleUsesSdk issue instead.

Test: Existing updated
Bug: Not filed
Change-Id: I7fb8d505322f5f4798bf7cc903bb549d7e619cce