Autoupdate of 2 submodules
*tools/adt/idea updated to 69d195e7769039aebb3eb6e51bc69b9bbd19f7d8
Registering AndroidLintChromeOsOnConfigurationChangedInspection

Addition of the lint rule that checks for finish() inside
onConfigurationChanged() to the AndroidLintBundle and enabling it by
default.

Bug: 236871629
Test: Auto generated
Change-Id: I6c469ba21a852e043be0810e96569ed14fa682f2

*tools/base updated to 874b299da760a9f30d4937a1ad1558cd06398485
Creation of Lint Test for identifying finish in onConfigurationChanged

ChromeOS has a resizeable VM where onConfigurationChanged() is called
when resizing, similar to swapping orientations. If finish() is called
inside, it can cause a bad user experience, and poor performance for
users on resizeable windows for large screens.

Bug: 236871629
Test: ChromeOsSourceDetectorTest#testFinishFoundInsideOnConfiguration
Change-Id: I85aa5d0191c6e47d62f5b95b06ce77ff10b1f3d2