Merged manifest view: Fix dynamic feature manifest view
The merged manifest view maintains two lists of dependency manifest
files, 'myFiles' for active contributors, and 'myOtherFiles' for
dependencies that did not contribute (or all their contributions were
overridden).

The previous code added things ad-hoc to the contributors list
when building the UI, which was removed in
https://cs.android.com/android-studio/platform/tools/adt/idea/+/1d3d27a18d83571c92d7e25bcffa76a2cc765c9c

However, it appears that recordLocationReferences misses some
references. For now, just index both myFiles and myOtherFiles when
choosing highlight colors, and return a positive index even if neither
list contains the file pending investigation of
recordLocationReferences.

Fixes: 247049256
Test: Partially covered by new integration test, although doesn't excercise the UI code.
Change-Id: I9e6be3efb15b9d28b81ea8466158b0de0d53e9ef