Adapt to changes in font in Build Variants tool window
Where we override the Table Cell RendererComponent completely (by
returning our own JComponent rather than by invoking the super method
of getTableCellRendererComponent()) we need to be a bit more careful
about setting it up appropriately, as it is re-used for subsequent
drawing operations, but the component is not part of the hierarchy.
If UI Settings (such as the global font) change, the table will be
repainted but we must be careful to propagate changes which the table
has picked up to our own renderer component.

Bug: 246717675
Test: manual
Change-Id: Ifd1848b8d473ab33dcf0d02ec2d3ba8f1c7468dc