Autoupdate of 2 submodules
*tools/adt/idea updated to 619e66b6012296fe86c1b79be2c03da54689e2df
Use drop-in library instead of experimental time API

Few other random trivial cleanups as I was replacing this API.

Bug: 254530383
Test: Existing
Change-Id: I590c33db6e2134c7cc30e1de0dff4a6502b40257

*tools/base updated to f6ded82fcaed3d1d138c174406341f729677cd8d
Create drop-in replacement for unstable Kotlin Time API.

This API is experimental meaning the underlying implementation may
change, breaking existing code. This happens to us because our
runtime Kotlin may be different than our compile-time Kotlin.

This drop-in library attempts to provide basic compatible
functionality so that at a later date we can drop in the replacement
when it no longer requires the risk of an experimental API.

Bug: 254530383
Test: Included
Change-Id: Ib7c1bc9ca79b67370801c334975a1db8e7dbaf9a