Public the getSshTunnelMonitor method in RemoteAndroidVirtualDevice class
In multi-device boot test, we need to access each device's SshTunnelMonitor from the test logic (outside of the RemoteAndroidDevice class) to validate the behavior. Previously it was not needed because for the single device boot test, the only SshTunnelMonitor reference is stored as a member variable in the test class.

BUG: 231337999

Test: Build and unit test
Change-Id: Ibb0012b1ccf7919631ad754c33412f5797699832