Add power data to to system trace data request
To fetch and handle the power data being traced
as part of the system trace, an additional request
has been added to the list of requests.

The handler for this request adds the power data just
fetched to the TraceProcessModel. Therefore the models
and respective adapters for trace processing have been
updated accordingly.

Bug: 254116554
Test: added and updated
Change-Id: I58e24cf0b76a5529d977dc32291545fe4980917c