Skeleton filter service for TSDKV upgrade assistant
Creates a project-level Light Service for executing filters.

Filters are functions from a Project to a List of some kind of
LocationProvider (currently PsiElement).

The FilterService keeps a map from String keys to Filters.
The key will eventually be the argument verify buttons pass
to indiciate which filter they're associated with.

The FilterService provides one method: the ability to call a specified
filter and get back the results. This would be called by the compiler
(to figure out which steps are relevant) or by verify action buttons.

An example "always irrelevant" Filter is included in the map.

I tested out that the Light Service style (using @Service instead of
declaring the service in the plugin.xml) works by adding a hello
function that printed to stdout and calling it from the bundle creator.

Not marking the "create a filter service" task as fixed yet as this will
probably require some revision.

Bug: 247643780
Test: N/A
Change-Id: Id6e9a360296c81a41cd5223761ee3e1533d2dff6