Fix Fast Preview detection in ComposePluginIrGenerationExtension
ComposePluginIrGenerationExtension was not correctly detecting it was
being invoked under Live Edit/Fast Preview and logging the
IncompatibleComposeRuntimeVersionException to stdout.
EmbeddedCompilerClientImplTest would pass since it does not verify that
the Compose transformations are applied, just that the compiler is
indeed being invoked.

After the fix, EmbeddedCompilerClientImplTest would fail because of the
incompatibility exception. This change also extracts out the code to
load the Compose runtime into the classpath allowing the compilation to
proceed as it does for Live Edit tests.

Fixes: 254505040
Test: EmbeddedCompilerClientImplTest BasicCompileTest
Change-Id: I9637cad40b5544ef370670c7432bd763f0f52ad4