Prevent Run Tool Window from being activated when profiling
The "activate tool window" checkbox in Run config had been ignored
in Studio for a very long time until recently. But fixing that
accidentally causes the Profiler tool window to be taken over by the
Run tool window.

This is a temporary fix to unblock profiler users until we can find
a better fix. The side effect is when user launches profilers, the
"activate tool window" checkbox will be unchecked regardless of
what it was before.

Bug: 251297822
Fixes: 254312239
Fixes: 254311743
Test: manually
Change-Id: I1c14a21fefb16bd9848482978336bd487ef854fe