Autoupdate of 2 submodules
*tools/adt/idea updated to a628f9782064a28c1030cfed20c5c46dc8b218ba
Support for SDK extensions in lint's API checker

This CL provides support for changes in lint to support mainline
modules. Most of the work is in lint; this CL mainly updates API
references that were changed, adds a few IDE integration tests, and
updates the quickfix to handle inserting SDK extension lookup checks.

Test: AndroidLintTest
Change-Id: I36e54cd344e5f71fcc90c9aed447b6d14e38595d

*tools/base updated to e4d9693071a1757baa9d431ac9b38926e259561c
Support for SDK extensions in lint's API checker

Design doc: go/lint-sdk-extensions

This CL updates lint's API check and associated features to be
extension-SDK aware. In the past, each API had an associated API level
when it was introduced. However, with extension SDKs, APIS can now be
backported to earlier versions, and so each API actually has a *vector*
of API levels, for the various SDKs it is available in.

This required extensive changes to the API check, as well as the version
checking utility. We already had an ApiConstraints class which is
recording API ranges for the platform SDK; it has now been extended to
be associated with a specific SDK, as well as to be able to carry
multiple dimensions of SDKs.

In terms of features, API check will now check compatibility across
these API vectors, and it also respects the new extension check utility
methods (and inserts these as quickfixes for API violations).

This CL also improves the version checking handling for a number of
scenarios; see for example VersionChecksTest' testWhenFallthrough,
testUncertainOr, testKotlinWhenStatement2

I ran lint against AndroidX with these changes and verified that the
baselines are mostly not affected; a few warnings were removed as a
result of more correct version checks.

Remaining work:
* Gradle DSL needs to include SDK extensions and for this to populate
  lint's minSdkVersions. Right now it just reads these values from
  the merged manifest.
* The API parser does not look for API vectors on super classes and
  interfaces; it only does this logic for the since= attributes
  (SDK 0).

Test: Existing, and many, many new tests included
Change-Id: I7bacc3174a787cb64b1b1642cd5e0cc2a04c7441