api-versions.xml: rename from -> sdks
Rename the `from` attribute of <class>, <method> and <field> tags to
`sdks` to better reflect that it represents a list of SDKs and
corresponding versions.

Historical note: `from` was originally chosen as a synonym for `since`
which the new attribute supersedes, which will be a somewhat obscure
reference once `since` is eventually removed.

Bug: 257175493
Test: ./gradlew test
Change-Id: I06ea47ad0a4e67b1e7ee1b5633adfbdaf8eceb5d