update-api updates less timestamps files
Before: rm then cp
Now: rsync

Removing files then copying them again causes too many things to
be changed. I considered only removing the files that wouldn't be
copied and then using 'cp -u', but rsync already does all this.

In order to discover this issue, first I tried to update APIs, then
I realized it took a lot of time. Then I asked for help! and ccross
told me:

    SOONG_UI_NINJA_ARGS="-d explain"

ooooooooh :) nice! So I put this in, and what I found was this:

    ninja explain: out/soong/globs/build/79 is dirty
    ninja explain: restat of output out/soong/globs/build/510 older than most recent input hardware/interfaces/graphics/common/aidl/aidl_api/android.hardware.graphics.common/current (1667499986272802243 vs 1667600937794675698)
    ninja explain: out/soong/globs/build/510 is dirty
    ninja explain: restat of output out/soong/globs/build/999 older than most recent input hardware/interfaces/graphics/common/aidl/aidl_api/android.hardware.graphics.common/current (1667499986268801917 vs 1667600937794675698)
    ninja explain: out/soong/globs/build/999 is dirty

oooh! dirty globs! dirty dirty globs! have some rsync.

Btw - this is still not so great, because the build system needs
to be invoked and rebuild the ninja file twice in some cases. This
is a reason why we should move towards using scripts for update-api.

Fixes: 257277612
Test: the following procedure, after the change
    m android.hardware.graphics.common-update-api # slow, b/c long build
    m android.hardware.graphics.common-update-api # fast, b/c no files changed
    # ... modify .aidl file ...
    m android.hardware.graphics.common-update-api # fast, b/c only file change (this is this fix)
    # ... add new .aidl file ...
    m android.hardware.graphics.common-update-api # slow (because of new file)
    m android.hardware.graphics.common-update-api # slow (because of new file in aidl_api)
    m android.hardware.graphics.common-update-api # fast, b/c no files changed
    # ... modify .aidl file ...
    m android.hardware.graphics.common-update-api # fast, b/c only file change (this is this fix)
    # ... delete .aidl file ...
    m android.hardware.graphics.common-update-api # slow (because of removed file)
    # visually check rsync '--delete-before' is reflected
    m android.hardware.graphics.common-update-api # slow (because of removed file in aidl_api)
    m android.hardware.graphics.common-update-api # fast, b/c no files changed

Change-Id: I0f7153dc6393463694215fe81ca47a9718d77ac0