Address issue with quitting too soon
The full details are in the bugs and the comment added in this commit.
The short story is that we sometimes try quitting Android Studio before
initialization is done, so this workaround is to wait a minimum of 10
seconds before quitting. For most tests, this will result in no
difference in behavior since most end-to-end tests have Android Studio
open for much longer than 10 seconds before being able to perform their
verifications.

Bug: 253431062
Bug: 256687010
Test: StartUpTest and UpdateTest should now work
Change-Id: I6b53651ebbe3b0d95e4189b66172bade9e1edcbf