Replace UserOptions in ProfilingConfig's fromProto()
The ProfilingConfiguration class has two utility methods
used to convert a UserOptions to a ProfilingConfiguration
(fromProto) and from a ProfilingConfiguration to a
UserOptions proto (toProto).

This change prepares the studio-side code for the removal
of UserOptions in favor for a more robust TraceConfiguration
proto as the main source of user-configurable options.

Bug: none
Test: updated
Change-Id: I1d24166b1a032aa390fe96479d313494cf6b7e67