Autoupdate of 2 submodules
*tools/data-binding updated to 69aab8c10428ff5664c10af44e5b6ed67b28a769
stop using a deprecated API.

The extensions-support was still using a deprecated API
that never graduated to stable. Replaced it with the newer
API that is now stable.

Test: existing
Bug: N/A
Change-Id: I9198cf772ed0a63fd6649435d7a882da3dd3a8fc

*tools/base updated to 8e496713acce67efd36916dfc446333eb12ad4f3
removed deprecated ALL_CLASSES_DIRS and ALL_CLASSES_JARS

API never went stable and was deprecated in 7.4

the gmaven-aars.txt needed to be modified as the code used to
generate the databinding extensions library was changed to accomodate
the API removal (see CL in the topic).

Test: Existing
Bug: N/A
Change-Id: I06d2c2c4e811b4462e456ae8a7ca6cd88e556f90