Autoupdate of 2 submodules
*tools/analytics-library updated to 1ebec7abc203909b2c0074861e32775f8ecf91a2
Analytics for the new "Add" Artifact API

Bug: NA
Test: Existing tests
Change-Id: I2f0d60aaca1d970927e6f61b959b6debe56e0d81

*tools/base updated to af4f438d8473cfb0121e6e71df2ce2a8ea82b22d
Add Artifact API to include external debug symbols

This change add a new API to add a new artifact to a multiple artifact type. With this API an existing directory with debug symbols can be added to a newly added artifact type. MergeNativeDebugMetadataTask consumes this artifact to get all the directories with externally sourced debug symbols. These debug symbols are merged with the debug symbols extracted from the unstripped native libraries.

Bug: 208593320
Test: ArtifactsImplTest, MergeNativeDebugMetadataTask
Change-Id: Icf30d403d1d4024bb71eabb9058a73b6447ac0bb