Add ability to send in additonal options in addOptions
The addOptions method noe takes in an "additionalOptions"
parameters, which is essentially a property bag. This was
done as the start up tracing does not have the profilers
instantiated in time to get access to profiler specific
data, so there needed to be a way for non-startup and
startup tracing to get the needed config info at runtime.
Each type of profiling configuration will use the property
bag differently depending on what options it needs added.

To constrain the use of the property bag/map to only access
certain fields, an enum of AdditonalOptions was defined to
constrain the key set.

Bug: 258515754
Test: added
Change-Id: I11287c9206194d8794798757cf5d78e707ea640d