Fix AndroidWearConfigurationEditor class lookup
The inheritance lookup is broken for Kotlin classes since the
inheritance index seems to be broken causing the parent classes to not
be found when using a module scope.

This change is a workaround where we use a ProjectScope.all for the
initial search and then filter down based on the correct scope.

Fixes: 251795458
Test: Manually verified since there is no practical way to generate
parent classes that are not physical in tests.
Change-Id: I2745e116324cef7dd9c317edcf923e0847aee348

Change-Id: I0c49595cb337deeb2ff68023443f811613ca2a4b