Autoupdate of 2 submodules
*tools/adt/idea updated to 2bdaf8f2e47ad6f42b8bf05fbbcd73f661efaf05
[AUA] Convert remaining uses of GradleVersion

All remaining uses of sdk-common's GradleVersion in the AGP Upgrade
assistant are for generic versions: versions of AGP are all
represented by AgpVersion, and versions of Gradle itself are
represented by Gradle's GradleVersion.  Replace all the remaining uses
with the equivalent sdk-common Version.

Bug: 242691473
Test: existing
Change-Id: Id0da319a97df52c4e7a05134f35b2d7435d0eeaa

*tools/base updated to 1629a3293f87e9b095724a3298edd3de4da980c4
Implement Version prefixInfimum factory

It turns out that we need this to support relatively simple semantic
questions, such as "is this version's leading part numeric?" which is
equivalent to "is this version greater than the prefix infimum of
`0`?".  This question is asked in various places, particularly when
parsing something which may or may not be a version (e.g. it might be
a string interpolation incautiously passed to be parsed as a version).

Uglify the comparison and equality code in Version substantially in
order to handle prefix infima.  Make it less ugly than it could be by
using the trick of computing an effective parts list one bigger than
the parts list of its counterpart, to ensure that explicit sequences
of DEV are correctly handled.

(Multiple-dispatch would make this substantially clearer.  Alas.)

Bug: 242691473
Test: added
Change-Id: If2498f9d32314fb9fc089b50ab91c440563da9bc