[core] Do not copy stream if all bindings are constructor args
When deserializing an entity which is instantiated via a constructor
with parameters in the general case we need to make a copy of the
input stream so that the constructor arguments can eb deserialized
first. However, when all the serialized values are constructor arguments
we do not need to make a copy as it is slow and inefficient.

closes https://github.com/JetBrains/intellij-community/pull/2232

(cherry picked from commit 00fb3e1c312eaec766e9509264422eb75d374a9d)

Change-Id: I55353cd59bde6d444ac4f718e79744adb669002d