Replace proto TraceType with idea side enum
The UserOptions.TraceType proto is used quite
frequently on studio-side to either help construct
a UserOptions proto, or to analyze profiling configs
and trace results.

As we move towards removing UserOptions, we will inherently
remove the child field of TraceType. To facililate this
removal, this change replaces the proto-based TraceType
with a studio-side enum (except some few places where
it is too tightly coupled with UserOptions). This preserves
the functionality, while taking a large step in replacing the
UserOptions proto reliant code on studio-side.

The aforementioned instances of UserOptions.TraceType that
still exist will be removed once futher changes are made to
completely remove UserOptions.

Bug: 258515754
Test: updated
Change-Id: I7a69719b4e1373c6392582f5a1be1c10e18f3c41