248054901: Fix handling of test and generated sources for project.xml files
This CL updates the UAST source list initialization such that it can be
overridden by a project implementation. For the project.xml-backed
project, where we have accurate information of whether individual files
are tests or generated (or both), use this information to directly
populate the UAST lists. (Prior to this we were inferring source roots
from the files (which we still have to do, for other reasons), and then
remembering whether these roots were generated source roots, test source
roots plain roots, and then we were classifying individual files based
on which root they were within. This doesn't work when you have
overlapping folders (as in the scenario from 248054901), and we don't
need to do it that way since we have direct info.

(Also fix some warnings while touching the code.)

Fixes: 248054901
Test: Unit tests included
Change-Id: I082ce523b2003b0deb3323406d04bb73867c96dd