Determine if color resource is file-based or not
RenameResourceReference uses the concept of 'isFileBased' to decide what
names and characters are valid when performing a refactor for resources.
Color and Drawable can be either, but are currently always listed as
file based. This leads to b/259111489, where some valid characters for
non-file-based color resources are incorrectly disallowed.

I'm moving the `isFileBased` function into RenameResourceReference.kt,
as that's the only place that it is used, and since with this change its
logic is becoming even more tied to this specific use case.

Fixes: 259111489
Test: RenameResourceReferenceTest
Change-Id: I23d3f12669bf96d77de60ad4533462da5a1571f8