VINTF treble test: clear version error fallout
In HIDL, if someone declares the wrong version in the manifest, this
is caught when we check the VINTF manifest at runtime, but because
AIDL minor versions were added later, and they aren't checked by
servicemanager, we rely on this test. So, it happens to be the case
that people frequently hit this error in vts_treble_vintf_vendor_test.

When they raise the AIDL version, but they don't raise the manifest
version, this causes an additional error to be surfaced here saying
the hash is notfrozen and that's a problem for a released interface,
but it's a red herring.

Fixes: 259589390
Test: vts_treble_vintf_vendor_test
Change-Id: I159c9c4a63a9cff581b57ea8b94ed772092160dd