[AUA] Fix deadlock
It's possible for the Upgrade Assistant to request a smart-mode read
action while under a modal progress dialog.  If this happens and
indexes happen not to be ready at that point, the parsing and cacheing
of information will deadlock (in a "soft" way, because the progress is
cancelable).

Make sure that the system is in smart mode before starting the modal
progress display.

Bug: 260151088
Test: manual
Change-Id: I011f7173eba3914a54fe7e055921a2d11f1ef4e1