[AUA] Custom user selector ui infra for the processors
This change makes possible to add custom ui elements for the
processors that require additional user input to proceed.
Previously we only had option to add combobox selector, but this
change makes possible to add any complex selector required, e.g.
jdk selector.
This change does not add any new code on the execution path so no
automatic tests are added here. We can not add a test until we
implement a step that uses this infra, so new test must be added
in the follow-up change. Was tested manually using
I3eee432f6461b5f569701a6faaffac09e51cb126 which contains an example
of using this API.

Bug: 251263800
Test: Will be added in next changes once we start using this
Change-Id: Idf9a0604e206340bde1a399ccfeb1525e6b42b62