Add the latest iteration of the extra large benchmark project
This project includes multiple diff files that alter the settings.gradle
files to alter the size of the project at will. This will be used to
create benchmarks with different sizes, used for different scenarios.

Source code is directly from: https://github.com/gavra0/android-extra-large-benchmark/tree/2022.9

Repository zip is created using the build_repo.sh but I had to generate
a separate project with a single module with all the dependencies for
the entire project, it was quite a manual process.

Test: N/A
Bug: 260381662
Change-Id: Ib13d3742fe89ad3d8461d42e3c113ab3ce2c965d