Fix custom option selection not reflecting
Per b/260201168, the change in selection of the custom
option dropdown reflected in the UI, but when a user
started recording, it was not using the selected option.

This has been fixed by overriding the setSelectedItem
method of the dropdown combo box to check if the user
has custom options selected, and if so, set the overall
selected trace config to be the one just selected in
the dropdown.

Fixes: 260201168
Test: added
Change-Id: Iaf26c42b4568956f6ad38f181ca6172050d3fd16