Live Edit: Minor cleanup of inlining analysis.
We are keeping track of binding contexts that we don't use because
we are only analyzing one file.

We are reanalyzing code that we already analyzed. That is probably
cached. However, since we just analyzed the code we
shouldn't have to do it again. We should just inspect the already
known analysis results in the binding context. When caches in
the resolution facade are flushed, that could avoid actual
reanalysis.

Bug: None.
Test: Existing inline analysis tests.
Change-Id: I1d3c6d8edfaf1e01da4c9b3020ab8cce36e2449d