Autoupdate of 2 submodules
*tools/adt/idea updated to a1c672c111d467b0ef9110465ad81be9beb35b9d
Remove UserOptions from start cpu trace commands

In ag/20587381, the base side has enabled the unified
trace config for the StartCapture method when handling
the START_CPU_TRACE command. Because of this, we can
now remove the UserOptions from the TraceConfiguration
created on studio side that we send to the base side.

Bug: 258515754
Test: updated
Change-Id: I7aa19d3f2770d63c483e288a72292dd154542a86

*tools/base updated to f029404863df43966c2d7b3fe40ec50d1b2fbae6
Enable unified trace config for start cpu trace

In an earlier change (ag/20535532) the adaptation to
the new unified trace config proto was made in StartCapture
of the trace manager. The use of the old vs new config proto
was controlled by a boolean flag passed in as a param to the
StartCapture method. After thorough automatic testing done in
that change, more manual testing was done with pre-O and O+
devices to make sure the trace manager was handling starting
a capture correctly as with the old config proto.

After verifying everything is still working as expected, the
flag to use the new unified trace config has been set to true
in the call of StartCapture made when handling the StartCpuTrace
command. The tests have been updated using the new unified trace
config as well to reflect this change.

Bug: 258515754
Test: updated
Change-Id: I84acef5cb4dd5f4881b68e78740d282d27a1cb66