Autoupdate of 2 submodules
*tools/adt/idea updated to 8efa20e8039f76795b4f3c2981c113d4c23ac978
Create AvdManager and DeviceManager caches

Define IdeAvdManagers and IdeDeviceManagers to manage creation of
AvdManager and DeviceManager instances within Studio, so that we share
DeviceManager instances.

Bug: 257226167
Test: updated
Change-Id: I0924b18bed6049a0caeccb8d6471f92c3d9f4fec

*tools/base updated to 09842e71fcfb5dba4ee8a5ad5a89c504220ff422
Remove caching from AvdManager

Rather than having unavoidable caching at the lowest layer, allow
clients to control the lifecycle.

Also support supplying a DeviceManager, so that it can be shared with
other clients and not get out of sync.

Bug: 257226167
Test: existing
Change-Id: I710c9fd84dea8e1a67a330bee677dbdda1aea2b0