[AUA] Make AllDone state take priority over warnings
If the Upgrade Assistant considers there to be nothing to do, do not
display warnings about project files not being clean: that is firstly
quite a natural (intermediate) state, and also the state definition
assumes that running the processor is a valid thing to do at that
point.  Instead, have AllDone take priority.

Bug: 261007044
Test: existing
Change-Id: I9d90dcf068ce58820935bf2ad95dc0093cb1ab73