Use flags for JfrTypingLatencyReports
Overall usage of JFR for typing latency is controlled by a Studio flag
(which can be remotely changed by server flags if the need ever arises
to turn this feature off). There is also a config proto that can be used
to tweak the reporter's parameters and is controlled by server flags.

Bug: 259447928
Test: JfrTypingLatencyReportsTest
Change-Id: Ifbac43577b54afe0576a8358660674bbb74cb44b