Autoupdate of 2 submodules
*tools/adt/idea updated to 8d04cc09ead6331f06c16c67e9fa75caeda0b575
Remove top-level symbol dirs usages

In a complementary change the top-level
symbol directories field in the TraceConfiguration
proto is being removed, as only the SimpleperfOptions
proto needs such field. There was one usage of this
top-level field on studio-side left. It has now been
replaced with the SimpleperfOptions based field.

Bug: 258709263
Test: existing
Change-Id: Iced590b1d56358385a0ebe82cdd3d1a6216a1ffc

*tools/base updated to cacf9e7d743434921ec62e7ba2e537c159c19d87
Remove UserOptions definition & proto usages

The UserOptions proto is no longer needed after
being replaced with technology-specific options
in the TraceConfiguration proto. Thus, it has been
removed.

The CpuProfilingAppStopRequest has a field of
type UserOptions.TraceType. As this field is
part of UserOptions, this field was removed accordingly.

Bug: 258515754
Test: existing
Change-Id: I75676165793ff183a73936bf20af2e214bed81b2