[Ps] Make GradleWrapper utilities accept GradleVersion
When referring to a version of Gradle, mostly use the
org.gradle.util.GradleVersion class rather than strings.  This should
make it substantially harder to accidentally pass in something that
isn't a Gradle version to the utilities, as was accidentally done in
the AGP Upgrade Assistant tests (though not in production).

Leave updateDistributionUrlAndDisplayFailure() as accepting a String,
as it does its own error handling to display a notification to the
user; handle the IllegalArgumentExceptions from parsing the possibly
user-supplied string accordingly.

Existing tests are mostly directly converted, except that the actual
GradleVersion parser is more restrictive about what it considers a
legal Gradle version string: the snapshot suffix format is
specifically ({Alpha}+-\w+)?(\d{14}|SNAPSHOT), so adapt tests to match
this.  (This is consistent with the resources available at
https://services.gradle.org/distributions-snapshots/ )

Bug: none filed
Test: existing tests updated
Change-Id: Id95487009c2054fcaa1d7ac17cddf79f625a2234