Fix Visual lint tests using wrong configuration
RenderTestUtil.getConfiguration returns a configuration that is cached.
So when asking for a configuration with a new device will update the
device of the cached configuration. That means we can't store
configurations with different devices to be used later, as they are all
the same configuration, with the latest device set.
This makes sure we ask for the configuration with the correct device
every time we want to use it, and we do not store those configurations.

Bug: N/A
Test: test code
Change-Id: I086db17bdc35275a58a7fe2c85f2a16ceafee6b8