Autoupdate of 4 submodules
*tools/data-binding updated to b4743bea1b3d7422f6441769e71a0d8abcb74029
Separate version from coordinates in maven_library

In preparation for using a select, which cannot be resolved in scripts.

Bug: 259300391
Test: Included
Change-Id: I20b1daae4b2269babf9664665aed50c148ea0378

*tools/analytics-library updated to de1187273899461274d9088b2271e30ade829173
Separate version from coordinates in maven_library

In preparation for using a select, which cannot be resolved in scripts.

Bug: 259300391
Test: Included
Change-Id: I579344018b36417f9d41ca8e9a5b3f60d14eab66

*tools/apkzlib updated to f106d323effdbc8f2dc98967285f21d7c0466bf1
Separate version from coordinates in maven_library

In preparation for using a select, which cannot be resolved in scripts.

Bug: 259300391
Test: Included
Change-Id: Ibcdce6271f57a9d95d05de2b02303a683045abcb

*tools/base updated to 823796fd2397eab142ec7882878ccaef253ba290
Separate version from coordinates in maven_library

In preparation for using a select, which cannot be resolved in scripts.

Bug: 259300391
Test: Included
Change-Id: I2c6251ed31f62011ca2609e0e97c727950f2d6a3