Autoupdate of 2 submodules
*tools/base updated to bc97c9b4dd70066fe2d91f801070f5cdb5e7caba
Replace Activating and Deactivating states with field

Having separate Activating and Deactivating classes forces clients to
handle four states instead of two, when many do not need to distinguish
beyond Disconnected and Connected. We can instead represent this
information with the "isTransitioning" and "status" fields. The "status"
field also allows us to be more descriptive about what is going on.

Plugins may still subclass Disconnected and Connected if it is useful.

Bug: 241591222
Test: updated
Change-Id: Icf5f8dc93179c9fdf7d5f5f85633bf0ac3189638

*tools/studio/google/cloud/testing updated to 8de9b2ca964ffcb1f0867c72ed2bdd8324ab3b9f
Replace Activating and Deactivating states with field

Having separate Activating and Deactivating classes forces clients to
handle four states instead of two, when many do not need to distinguish
beyond Disconnected and Connected. We can instead represent this
information with the "isTransitioning" and "status" fields. The "status"
field also allows us to be more descriptive about what is going on.

Plugins may still subclass Disconnected and Connected if it is useful.

Bug: 241591222
Test: existing
Change-Id: I24eb615df7c035167127782bac8f978ca2ec65ef