Cache FilterService results for cheap reads
We don't want UI elements to have to execute a filter to determine
rendering. Instead, they can query the cached results.

The cache initially maps each filter to a `null` result and is updated
whenever a filter is called.

If/When we start executing filters in the background, we can update the
cache to a `null` result first to indicate a "result not ready yet"
state.

This can probably be cleaned up a lot, but that can be done later.

Fixes: 247652813
Test: included
Change-Id: I48135a899e98cba00f35545ae31a618219113600