Autoupdate of 2 submodules
*tools/adt/idea updated to fa378851ea83599ce79db1713a3ec27c551d8887
Update the model when a root is removed

When the screen is locked the agent can no longer retrieve view
information from the device. Update our view hierarchy to reflect
this.

Bug: 260106225
Test: Added unit test
Change-Id: Ibfbfac8d86f39eb2efdbf5459697635e1be49cba

*tools/base updated to 1a8cd3f21d36b147a1665231a792f95a833ae7af
Do not give up when there are no root views

Instead send an empty layout event.
The client needs a layout event to complete the attach progress.

Bug: 260106225
Test: Adapted ViewLayoutInspectorTest
Change-Id: Ia832874441e26bfbcd0b18452b99335a0c077669