Add @Used/@Unused annotations to track progress
This CL adds @Used and @Unused annotations to both com.sun.javadoc API
interfaces and to implementation classes – to track the overall progress of how
much is still to be implemented.

Both @Used and @Unused annotations have an "implemented" property to distinguish
between implemented and unimplemented states which are then used in
print-progress.sh script.

Current progress is 67/136 methods implemented (49.26%).

Bug: 260694901
Test: ./print-progress.sh
Test: ./gradlew doclava17:test
Change-Id: Ie01d9999aa6211bdcf09fd1520b621d5990bbc0c