Autoupdate of 2 submodules
*tools/adt/idea updated to a400a86a350b629b810bc1576afaf49969c51758
Split the DeviceHeadsUpListener.deviceNeedsAttention method into three methods

The separate methods will be used in a subsequent CL to differentiate behaviours of the Running Devices window in different situations.

Test: existing
Bug: N/A
Change-Id: Ic7e75b4bb1b11a56f4e90f570f7d0d869675435a

*tools/studio/google/cloud/testing updated to c43cfe0b83b8d46931c8d5aec6c3f498e5871da5
Split the DeviceHeadsUpListener.deviceNeedsAttention method into three methods

The separate methods will be used in a subsequent CL to differentiate behaviours of the Running Devices window in different situations.

Test: existing
Bug: N/A
Change-Id: Ic0870b054ed327ec763987e91c2db5eaa4b45ef2