Autoupdate of 2 submodules
*tools/adt/idea updated to 0f069cfca21271f60fb324ce30d5c20a26233fc5
Add missing flags needed to start Game Tools

In order to start game tools, two new options needed to be passed with
game tools:

  -Djava.system.class.loader=com.intellij.util.lang.PathClassLoader
  -Didea.vendor.name=Google

In order for PathClassLoader to work, we need to set the vendor name. If
we don't, start-up will fail with an invalid state exception as IntelliJ
will start another activity in an attempt to figure out the vendor name.

Bug: 263887432
Change-Id: I02277d9d3450b93e064d7d6b3f345fd33ba98938

*tools/vendor/google updated to 9be7c51fe9a67452bfb4303b947b601cbaa73e3c
Add missing flags needed to start Game Tools

In order to start game tools, two new options needed to be passed with
game tools:

  -Djava.system.class.loader=com.intellij.util.lang.PathClassLoader
  -Didea.vendor.name=Google

In order for PathClassLoader to work, we need to set the vendor name. If
we don't, start-up will fail with an invalid state exception as IntelliJ
will start another activity in an attempt to figure out the vendor name.

Bug: 263887432
Change-Id: I7c18545d6b8025a4473ac5d9a401ef49a1ded1fc