Autoupdate of 2 submodules
*tools/adt/idea updated to 18fa073b71fe7fc3998c46ca2b765fb11f860612
JDWPTracer: Generate output upon close

Currently traces are generated only when VM:Exit reply is detected
. This is an issue because this packet is not always issues in the
case of a debugger failing to attach.

The solution is to add a way for ADBLIB/DDMLIB to let JDWPTracer
know when a debug session has ended.

Test: NA
Bug: NA
Change-Id: If60106456fb050270dba6b3632827789dea5b2dc

*tools/base updated to defec3ee58f9c15756beb19f2412fd2fc67d66ca
JDWPTracer: Generate output upon close

Currently traces are generated only when VM:Exit reply is detected
. This is an issue because this packet is not always issues in the
case of a debugger failing to attach.

The solution is to add a way for ADBLIB/DDMLIB to let JDWPTracer
know when a debug session has ended.

Test: NA
Bug: NA
Change-Id: Ia1689c7766da49ddc3b0e72eb47b43c878470e49