Autoupdate of 2 submodules
*tools/adt/idea updated to f3d6042a997a5e04e77dede659f42c3fd7c534a6
Remove redundant Mockito cleanup code

MockitoThreadLocalsCleaner.cleanupAndTearDown() already clears all
Mockito thread-locals on all threads, so there is no need to attempt
to do the same via resetOngoingStubbing(). The latter was less
thorough anyway, because it only affected app pool threads instead
of all threads.

Bug: 263278194
Test: existing
Change-Id: I5fa80ae6c7b90e4e008f335814019b86f38eaea2

*tools/base updated to cf7324b98864e1537062e438248d53578ce32be1
Remove unnecessary Mockito cleanup code

The call to Mockito.reset(mock) is redundant with
resetWellKnownThreadLocals(). Furthermore, Mockito.reset() only has
the intended effect if it is called on the same thread on which the
mock was created---which we cannot guarantee in general.

Test: existing
Bug: 263278194
Change-Id: I055ef4a9dc7b3d40ec12060d3006f9fddc2a7d1a