Autoupdate of 2 submodules
*tools/adt/idea updated to 8f5246363e00ddbf3a167de3f759113eea1a9378
LE: Don't send over method desc on request

We no longer use the method's name and type on the device to
determine how recompostion is done. Instead we are going to
rely solely on sending the recompose group ID to trigger
recomposition.

Since there is now no need to determine the final signature
of a composable method, there is no need to make sure of the
Kotlin TypeMapper that is becoming invalid in the new IR.

Finally, this CL focuses on purely removing code for easier
reviewing. I'll be send out follow up CLs to future simplify
the communication protocol.

Fixes: 260327461
Test: Existing

Change-Id: I23b941b698a40532670632313593f633166c683b

*tools/base updated to ac0a3a41cd4dcace7cc838c240cdf96406ad4f72
LE: Remove the unused map of Live Edited methods.

We are committing to using INTERPRET_ALL approach which
makes the map of which methods to interpret useless.

Bug: 260327461
Test: Existing

Change-Id: Ia32ec55b5976751a372d762e97f75fa56c1c312d