Autoupdate of 2 submodules
*tools/adt/idea updated to c436bdc3a5572e4fa2c9b10edf8c7a146ab91ecc
Fix StudioDefaultJDKTest by using expected JDKComboBox item naming

Due to the b/255545053 issue, we moved away from hardcoded jdk.table
naming, for this reason, we replace "Android Studio java home" with
JAVA_HOME following IntelliJ logic, but most important this ComboBox
is just a reference that internally is translated to #JAVA_HOME
macro, without creating an unnecessary jdk.table entry for the
user environment variable.

Bug: 263797479
Test: StudioDefaultJDKTest
Change-Id: I50a3719d0ec073daf655980b69e949e0d5cd05fc

*tools/swing-testing updated to 4304aa5ff6a7e2d25ca257f377a321498e63f142
Add support for different SdkListItem on BasicJComboBoxCellReader

IntelliJ supports different types of items as part of JComboBox,
described on SdkListItem. We recently started to use SdkReferenceItem
allowing us to don't create a linked jdk.table entry, something
fundamental in the case of macros like #JAVA_HOME. This allowed us to
move away from the hardcoded jdk naming we had for the same purpose.

This change allows us in those cases to be able to return properly
the value of the cell, instead of avoiding processing SdkListItem
as text, and unable to return its value.

Bug: 263797479
Test: StudioDefaultJDKTest and existing
Change-Id: I39e3da009eafb424c3afb1b659bb1dc7d75de62b