Autoupdate of 2 submodules
*tools/adt/idea updated to 3f82d61e85e910790ca26f5fe845d7e2065dfe9c
[li-overlay] Handle optional screenshots from agent

Screenshots capturing can be turned off. In this case events
from the agent are going to have no screenshot field.

This change updates ViewAndroidWindow to be able to handle
this case.

Test: AppInspectionTreeLoaderTest
Bug: 265150325
Change-Id: Id88baed9b158f446fc33b0909cc26448df568df4

*tools/base updated to 7591adbb12670e3d206d0ff01b188c88cfcdb6e6
[li-overlay] Make LayoutEvent#screenshot optional

This change is in preparation of the integration between
LayoutInspector and RunningDevices.
We won't need bitmap screenshots after the integration, but there
is going to be a transition period where both old and new
LayoutInspector will need to work. Which will require the
ability to switch bitmap collection on and off.

This change adds a `disableBitmapScreenshot` property to the
InspectorState.

When set to true, this property prevents the agent from capturing
and sending bitmap screenshot to Studio.

The change does not affect SKP images, as we will still need them
to capture snapshots.

The property can be set from Studio using
DisableBitmapScreenshotCommand.

Test: ViewLayoutInspectorTest
Bug: 265150325
Change-Id: I6c7fa6ffeca8810c0ddbe69891329dff5029ddb6