Disable IR inlining phase for Live Edit compilations.
Because of a change in the Kotlin 1.8 Compiler, it now ignores the Live
Edit request of not inlining final constant values.
This used to be controlled by using the LanguageFeature.InlineConstVals
in LiveEditCompilerLanguageSettings.

This causes the compiler to inline the light R class final value. Since
this value is "fake" and does not map to the actual resource, this makes
Live Edit and Fast preview fail when accessing resources.

https://youtrack.jetbrains.com/issue/KT-55892 has been reported upstream
to fix the compiler setting that allowed us not inline the final
constants.

Fixes: 265119058
Test: FastPreviewManagerGradleTest
Change-Id: I16c1a152217c6784575404b458bc3fb28b5dfb73