ddmlib: Delegate `getClientName` if needed
If a `DeviceClientManager` is active, the `DeviceImpl` delegates
looking up for a process name to it, instead of relaying of
its internal `mClientInfo` map. (This is similar to what was
previously done with the `mClients` map).

The reason this is needed is because the way for `mClientInfo`
map is updated with calls to `addClientInfo`, etc. which are
performed from the `ClientImpl` class, thus created a tight
dependency between `DeviceImpl` and `ClientImpl` that can't
be (easily) replicated when a `DeviceClientMAnager` is active.

Bug: 188067646
Test: Existing
Change-Id: I3fe10a06bdce1079b835923204b00522e0a8e0ab