Autoupdate of 2 submodules
*tools/adt/idea updated to ef30e9495941a600c2208d3670cb3d10c50bd86d
Make the power profiler feature flag dynamic

The feature flag for displaying the power rail and battery
counter tracks was made to be an enum/integer to be more dynamic.
This now allows us to disable (PowerProfilerDisplayMode.HIDE) the
viewing of the power and battery tracks, but allows us to add different
visualizations via non-zero positive integers mapping to a different
visualization query.

Bug: 254116554
Test: updated
Change-Id: I3a656229b12ee030d0299c063781d20e3ad6a098

*tools/base updated to 7913e73a122ff0a234991203684307a184892bcd
Adapt to dynamic power track feature flag

On studio/idea side, the feature flag for displaying
the power rail and battery counter tracks was made to
be an enum/integer to be more dynamic (the positive integers
are mapped to enums on studio-side). This now allows us
to disable the viewing of the power and battery tracks,
but allows us to add different visualizations via non-zero
positive integers mapping to a different visualization query.

This change adapts the trace processor daemon and related
protos to accept and utilize the aforementioned passed in
feature flag.

Bug: 254116554
Test: existing
Change-Id: I62a95eeb5e38472e8aa00adb23bcf2c8e06a660f