Autoupdate of 2 submodules
*hardware/interfaces updated to 2524ba06a2ee8936f4a893876e0ba5cba0039110
Merge "Adjust for library rename"
*tools/security updated to 498c61b3ad2ee6cfcd0ca27bbb16937bb4905a4b
Rename from BCC to DICE chain

Fully deprecate the term 'BCC' in this code base. Reexport the DICE
items to hide the internal structure of the module.

Move public key CBOR logic out of the dice module to match the structure
of the top-level module.

Test: atest libhwtrust_tests
Change-Id: Ie98e5e52587ec1786a3ca110c19b00d0831758ea