Autoupdate of 2 submodules
*tools/adt/idea updated to 4dbac97573e26b862171365f7157121f8195e978
Broaden version catalog support to TOML support

This CL adds TOML support to lint when running in the IDE;
lint will now run on the fly when you edit TOML files, the
suppress quickfix will insert suitable suppression comments
when you suppress warnings in TOML files, etc.

This also tracks upstream lint changes to handle TOML, where
it does its own .toml file parsing; this mean we no longer need
the IDE-side visitor.

Test: Existing and a few new unit tests
Bug: 256171234
Change-Id: I2280d52ea3fb659e7aaa854181d569d40d4a427d

*tools/base updated to 1c3dbf99553f824419a1e73bd8011ca449d09fe0
Broaden version catalog support to TOML support

This CL reworks the new Gradle version catalog support a bit such that
TOML files are now a top-level supported language type for lint -- they
have their own scope that detectors can opt into, with a supported API
for visiting and querying the TOML data, etc.

This means that errors are reported directly in TOML files instead of
indirectly via Gradle files (and in a separate CL on this topic lint
also starts running on the fly in TOML files in the IDE).

This CL also adds a new TOML-related lint warning: don't mix and match
version catalogs and plain Gradle dependencies. There are quickfixes to
help with this; if an existing version catalog variable already
references this library, just switch to it, otherwise you get several
fix alternatives, including adding a new declaration for this library,
or if the library is already in the catalog but with a different
version, you also get the option of either just using that other
version, or changing the catalog version to this current version.

In order to support TOML by detectors, lint needed an API to refer to
TOML data (similar to how it has UAST for Java/Kotlin and DOM for XML);
this CL includes a custom parser for TOML which is now used, which
produces these new model classes. This meant we don't need to have the
bridge visitor to tomlj and IntelliJ TOML, so this CL drops those.

Test: Many new unit tests are included (and the previous ones updated a
  bit)
Bug: 256171234
Change-Id: I6463b3cd3312f79a698a7b264bbec3ecaff1dbcb