Check if current method needs a deopt after method exit callbacks
Method exit listeners could request a frame pop. These are handled only
in the interpreter so we need to deopt at the method exit to handle
these. We were only deoptiming at the method exit if the caller needs a
deoptimization and hence missing frame pops if the caller didn't need a
deoptimization.

Bug: 266175266
Test: art/testrunner.py -t 1955
Change-Id: I10589e10c6b160a75f828302837d97189a79d175