264799025: GradleDependency lint: false positive when using explicit @aar qualifier
Lint detectors are passed the raw source strings for arguments to
Groovy methods. The code to produce the actual string from this
was simplistic; it was basically just removing the surrounding quotes.

For Kotlin strings in KTS files this isn't enough; it wasn't handling
things like basic string concatenations and string escapes and raw
strings -- these are all unlikely but should be supported.

There was a second problem in the KTS support where it was calling the
wrong source literal mapping function; the one called is actually a
source pretty printer instead of returning the actual source, and
this one actually turns strings with substitutions (like "foo:${bar}")
into concatenations (e.g. "foo" + bar). In the case of strings ending
with a string, this would look like "group:artifact:${version}@aar",
which was turned into `"group:artifact" + version + "aar"`, and the
earlier strip-surrounding-quotes code would just turn this into
`group:artifact" + version + "aar` (since the string starts with and
ends with a quote), and from there the confusion led to several
downstream problems, such as 264799024 and 264799025.

Fixes: 264799024
Fixes: 264799025
Test: Unit test included
Change-Id: Idee193d419b5b0a581125ae9a7e2ad7a8d10cb5c