Set dark theme as default for wear configurations
We were using Theme.DeviceDefault.Light as the default theme for wear
devices preview in Android Studio. This was first added in 2016, but
since then wear devices have changed their default theme to dark. This
can be easily done by changing the default theme for wear in ThemeUtils
to Theme.DeviceDefault.

Bug: 233251842
Test: N/A. Manual by creating a wear preview in Android Studio.
Change-Id: I5acc79d1b4b434aadd3ded36f779c0cdf1bd58c2