Autoupdate of 2 submodules
*tools/base updated to 57a252e8f35fb490c966cb65f103e5ef11f20644
Create FakeClock utility class for testing

Some of our code uses the Java8 Time APIs. Unfortunately, Java
does not provide a fake implementation of Clock() and leaves it
up to developers to write their own. We have a couple but they
are spread out all over the place. This will ensure that we have
one more. Maybe also we can replace the others.

Bug: N/A
Test: FakeClockTest.kt
Change-Id: I4b62ab2b1957bed8ae952eaed7cf96f223607652

*tools/vendor/google updated to 323f38455df9c0999528fe1f1041670226f3a75f
Replace FakeClock with impl from tools/base

I recently created an implementation of a FakeClock in tools/base
for more general use, and I'm cleaning this up as it can use that
instance.

Bug: None
Test: No changes to tests other than subbing in this library
Change-Id: I26f8f7571f6c30f04323503fc864ab07ee751995