Autoupdate of 2 submodules
*tools/adt/idea updated to b636ec005a0b4b3ebbb4d0b06568baec5229d871
Remove reference to JDK on Android Platform SDKs

The goal is to decouple the SDKs from JDK, by avoiding to extend
the AndroidSdkType from JavaDependentSdkType, that allowed until now
to keep a reference to the JDK. This reference was stored in the
jdk.table under SdkAdditionalData, using a name of the linked JDK i.e
<additional jdk="Embedded JDK" sdk="android-32" />

This was causing many issues to keep aligned the gradle JDK with
the JDK linked to SDK, but also points to wrong JDK minimum version
and end up in some scenarios to configure new projects with this JDK.
At the same time those new SDKs were creating using jdk.table entry
"Android studio default JDK" when present, option that users only can
configure the first time they open studio.
In addition, there were different UIs allowing them to modify
the JDK linked to SDK, causing more confusion to deal with
gradle JDK, project JDK and JDK linked to SDK.

To archive it, we replace the JavaDependentSdkType with
SdkType, keeping consistent with for example KotlinSdkType.

Bug: 254402612
Test: Existing ones
Change-Id: If86a7fd7091bb194227f655a1311b590bda6ace0

*tools/vendor/google updated to 4de2dc07a9e19a05dc3047c66e014074735f8f4e
Update snapshots with Android SDKs removed reference to JDK

Bug: 254402612
Test: N/A
Change-Id: Iffca368cc3e2563f39624c02a41a604306c7e2fc