Use bitmap to maintain which signal is being handled
The current approach of using a flag for all signals restricts from
having nested loops of different types. For instance, if ART has SISEGV
and SIGBUS handlers installed and the former raises the latter, it
should be allowed but isn't currently.

This CL replaces the flag with a bitmap, one bit per signal. This way we
only skip the same signal type as being handled.

Bug: 160737021
Test: art/test/testrunner/testrunner.py --host -t 2045-uffd
with https://android-review.git.corp.google.com/c/platform/art/+/2411972

Change-Id: I25250e3f83f6258be64b83d9661c89cf2e26d2b5