Remove FirebaseLogTask
This was added back in 2016 with the intent of doing something with the
logs but was never actually completed.

All it does now is turn on VERBOSE logging for tags FA & FA-SVC, but it
doesn't actually turn them off, claiming that turning them on is a
`relatively harmless change`.

I have found some complaints on stackoverflow that these log tags cannot
be turned off:

https://stackoverflow.com/questions/37577663/turn-off-firebase-logcat-logging

I think the value of turning these tags to VERBOSE is minimal. The docs
cover how to turn them on:
https://firebase.google.com/docs/analytics/cpp/events#view_events_in_the_android_studio_debug_log

And the code actually logs this to logcat:

```
I/FA: To enable debug logging run: adb shell setprop log.tag.FA VERBOSE
```

The main motivation for removing this code is that it uses the old
Logcat service which is scheduled for removal. It is possible to use the
new Logcat service but since the logs are not used anyway, that seems
like a waste of time.

Bug: na
Test: na
Change-Id: I531179f30e6cf38f90a0be21b5dbcd73d4e9f7f1